# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Quinn AI Personal Assistant API
  description: |
    Quinn is an AI-powered personal assistant that helps manage emails, banking, 
    food planning, life tasks, and more. All endpoints require authentication via 
    AWS Cognito JWT tokens.
  version: 1.0.0
  contact:
    name: Quinn Support
    email: support@quinn.ai
servers:
  - url: https://8aqsagpkp6.execute-api.eu-west-2.amazonaws.com/dev
    description: Development server
  - url: https://api.quinn.ai
    description: Production server (future)
security:
  - bearerAuth: []
...
paths:
...
  /banking/link-token:
    post:
      tags: [Banking]
      summary: Create Plaid link token
      responses:
        "200":
          description: Link token created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  linkToken:
                    type: string
                  expiration:
                    type: string
                    format: date-time
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /banking/accounts:
    get:
      tags: [Banking]
      summary: Get linked bank accounts
      responses:
        "200":
          description: Accounts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankAccount"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Banking not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /banking/transactions:
    get:
      tags: [Banking]
      summary: Get transactions
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: accountId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Transactions retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Banking not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

  # Food
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito JWT token
...
  responses:
...
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: "Unauthorized"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: "Internal server error"
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        details:
          type: string
        code:
          type: string
...
    BankAccount:
      type: object
      properties:
        accountId:
          type: string
        name:
          type: string
        officialName:
          type: string
        type:
          type: string
        subtype:
          type: string
        mask:
          type: string
        currentBalance:
          type: number
        availableBalance:
          type: number
        currency:
          type: string
    Transaction:
      type: object
      properties:
        transactionId:
          type: string
        accountId:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date
        name:
          type: string
        merchantName:
          type: string
        category:
          type: array
          items:
            type: string
        pending:
          type: boolean
...